+++ src/app/data/poe-prices/schema/item-price-prediction.ts
@@ -1,7 +1,7 @@
 export interface ItemPricePredictionResponse {
   min: number
   max: number
-  currency: 'chaos' | 'exalt'
+  currency: 'chaos' | 'exalt' | 'divine'
   warning_msg: string
   error_msg: string
   pred_confidence_score: number