+++ src/app/data/poe/schema/poe-api.ts
@@ -38,6 +38,7 @@ export interface TradeItemsEntryFlags {
 export interface TradeLeaguesResult {
   id: string
   text: string
+  realm: string
 }
 
 export interface TradeStaticResult {
@@ -295,14 +296,6 @@ export interface UltimatumFilters {
   ultimatum_output?: FilterOption
 }
 
-export interface SentinelFilters {
-  sentinel_duration?: FilterValueOption
-  sentinel_empowerment_limit?: FilterValueOption
-  sentinel_empowerment?: FilterValueOption
-  sentinel_durability?: FilterValueOption
-  sentinel_max_durability?: FilterValueOption
-}
-
 export interface MiscFilters {
   quality?: FilterValueOption
   ilvl?: FilterValueOption
@@ -343,7 +336,6 @@ export interface Filters {
   map_filters?: FilterGroup<MapFilters>
   heist_filters?: FilterGroup<HeistFilters>
   ultimatum_filters?: FilterGroup<UltimatumFilters>
-  sentinel_filters?: FilterGroup<SentinelFilters>
   misc_filters?: FilterGroup<MiscFilters>
   trade_filters?: TradeFilterGroup
 }