+++ src/app/modules/evaluate/component/evaluate-settings/evaluate-settings.component.ts
@@ -34,10 +34,6 @@ export interface EvaluateUserSettings extends UserSettings {
   evaluateQueryDefaultHeistRequiredLevels: boolean
   evaluateQueryDefaultHeistContracts: boolean
   evaluateQueryDefaultHeistBlueprints: boolean
-  evaluateQueryDefaultSentinelCharges: boolean
-  evaluateQueryDefaultSentinelDuration: boolean
-  evaluateQueryDefaultSentinelEnemies: boolean
-  evaluateQueryDefaultSentinelEmpowerment: boolean
   evaluateQueryDefaultMiscs: boolean
   evaluateQueryDefaultType: boolean
   evaluateQueryDefaultAttack: boolean
@@ -123,7 +119,7 @@ export class EvaluateSettingsComponent implements UserSettingsComponent {
 
   public onCurrenciesValueChange(): void {
     if (this.settings.evaluateCurrencyIds.length <= 0) {
-      this.settings.evaluateCurrencyIds = ['chaos', 'exa']
+      this.settings.evaluateCurrencyIds = ['chaos', 'divine']
     }
   }