+++ src/app/shared/module/poe/component/item-frame-properties/item-frame-properties.component.ts
@@ -55,7 +55,6 @@ export class ItemFramePropertiesComponent implements OnInit {
       'ultimatum',
       'incursion',
       'heist',
-      'sentinel',
     ]
     this.isArmour = this.item.category?.startsWith(ItemCategory.Armour) || false
     this.isWeapon = this.item.category?.startsWith(ItemCategory.Weapon) || false