+++ src/app/shared/module/poe/service/item/query/item-search-filters-type.service.ts
@@ -134,7 +134,10 @@ export class ItemSearchFiltersTypeService implements ItemSearchFiltersService {
       case ItemCategory.GemActivegem:
       case ItemCategory.GemSupportGem:
       case ItemCategory.GemSupportGemplus:
+      // leaguestone
       case ItemCategory.Leaguestone:
+      // memoryline
+      case ItemCategory.MemoryLine:
       // currency
       case ItemCategory.Currency:
       case ItemCategory.CurrencyPiece:
@@ -165,11 +168,6 @@ export class ItemSearchFiltersTypeService implements ItemSearchFiltersService {
       case ItemCategory.HeistBlueprint:
       // expedition
       case ItemCategory.ExpeditionLogbook:
-      // sentinel
-      case ItemCategory.Sentinel:
-      case ItemCategory.SentinelStalker:
-      case ItemCategory.SentinelPandemonium:
-      case ItemCategory.SentinelApex:
         query.filters.type_filters.filters.category = {
           option: item.category,
         }
@@ -180,7 +178,9 @@ export class ItemSearchFiltersTypeService implements ItemSearchFiltersService {
           option: item.category,
         }
         if (item.rarity === ItemRarity.Unique || item.rarity === ItemRarity.UniqueRelic) {
+          query.term = undefined
           query.name = undefined
+          query.type = undefined
         }
         break
       // prophecy