+++ src/app/shared/module/poe/type/item.type.ts
@@ -99,6 +99,7 @@ export enum ItemCategory {
   MapScarab = 'map.scarab',
   Watchstone = 'watchstone',
   Leaguestone = 'leaguestone',
+  MemoryLine = 'memoryline',
   Prophecy = 'prophecy',
   Card = 'card',
   MonsterBeast = 'monster.beast',
@@ -112,10 +113,6 @@ export enum ItemCategory {
   HeistContract = 'heistmission.contract',
   HeistBlueprint = 'heistmission.blueprint',
   ExpeditionLogbook = 'logbook',
-  Sentinel = 'sentinel',
-  SentinelStalker = 'sentinel.stalker',
-  SentinelPandemonium = 'sentinel.pandemonium',
-  SentinelApex = 'sentinel.apex',
   Currency = 'currency',
   CurrencyPiece = 'currency.piece',
   CurrencyResonator = 'currency.resonator',
@@ -185,7 +182,6 @@ export interface ItemProperties {
   ultimatum?: ItemPropertiesUltimatum
   incursion?: ItemPropertiesIncursion
   heist?: ItemPropertiesHeist
-  sentinel?: ItemPropertiesSentinel
 }
 
 export interface ItemProperty {
@@ -307,14 +303,6 @@ export interface ItemPropertiesHeist {
   rewardRooms?: ItemValue
 }
 
-export interface ItemPropertiesSentinel {
-  duration?: ItemValueProperty
-  durability?: ItemValueProperty
-  maxDurability?: ItemValueProperty
-  empowerment?: ItemValueProperty
-  enemiesEmpowered?: ItemValueProperty
-}
-
 export enum HeistObjectiveValue {
   Moderate = 1,
   High = 2,
@@ -367,7 +355,6 @@ export enum ItemSection {
   Relic,
   Incursion,
   Heist,
-  Sentinel,
 }
 
 export interface ItemSectionParserService {