+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.8.20 (TBD)
+
+- Updated PoE Assets to support 3.19.2b (Kalandra League)
+- Added support for the 'Memory Line' item category
+- Added support for the 'Engraved Ultimatum' base item type
+- Added 'divine' currency to the list of possible currencies for poe price predictions
+- Replaced 'exa' with 'divine' in the list of default evaluation currency settings
+- Fixed an issue with the setting's league list (i.e. added a realm=pc filter)
+- Fixed an issue with 'Unmodifiable' evaluation
+- Fixed an issue with the bulk trade message regex
+- Fixed an issue with Metamorph Samples returning almost no results due to the item name being used as a search term
+
+## 0.8.19 (2022-08-21)
+
+- Updated item parser and query to remove sentinel mods
+
 ## 0.8.18 (2022-06-17)
 
 - Updated PoE Assets to support additional mods for 3.18.1